requirements.txt
openai>=1.51.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
